requirements.txt
apache-libcloud==2.4.0
asn1crypto==0.24.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
Django==2.2.1
django-allauth==0.38.0
django-bootstrap4==0.0.8
django-constance==2.4.0
django-environ==0.4.5
django-picklefield==2.0
django-recaptcha==2.0.2
gunicorn==20.0.4
idna==2.9
mysqlclient==1.4.6
ndg-httpsclient==0.5.1
oauthlib==3.0.1
Pillow==7.1.2
pusher==2.1.3
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
pyOpenSSL==19.0.0
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rollbar==0.15.0
setproctitle==1.1.10
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
